// src/modules/notifications/enums/notification-target-type.enum.ts
export enum NotificationTargetType {
  PRODUCT = 'product',
  USER_PROFILE = 'user_profile',
  ORDER = 'order',
  REWARD_RESULT = 'reward_result',
  WALLET_TRANSACTION = 'wallet_transaction',
  CONVERSATION = 'conversation',
  NEWS = 'news',
  ACCOUNT_STATUS = 'account_status',
  ACCOUNT_SECURITY = 'account_security',
  SELLER_APPLICATION = 'seller_application',
}